Refactor Enums Service calls in instance manager
In MsbmsSystInstanceMgr we were calling a number of MsbmsSystEnums
functions with the old call style where we had to specify the service
name in each call.

This is no longer necessary since that service was changed with per
process service designation.

Update MsbmsSystInstanceMgr code to reflect the change insofar as the
module is complete.