Refactor functional_type_internal_name references
Across multiple modules and the database, we fully reference functional
type internal names in a fully qualified way.  While I think our rule of
thumb that has us use complete names is generally helpful, I think for
internal_name qualifiers we can just use "name".  Display name, external
name, and any other names than internal should be fully qualified.  But
for our purposes in the code we should take "name" to mean internal_name
unless there is a contextual reason to think otherwise.

I can't give a really good reason for this other than it will reduce
some typing and screen space consumption for what will be the more
common reference in logic; a decent rationalization is that the
application code is de facto the "internal" usage being discussed in the
name and so is redundant to include in the code that is the primary
consumer of the internal name.