Refactor instance type management to its own file
The Impl.Instances file was starting to get a little long and it was
starting to address multiple concerns.

We'll create a new module for InstanceTypes and split that instance type
specific logic into its own module and file.  We'll also do the same for
other similar distinct functionality groups within a single component.

This should help us be a bit more organized.