Fix hardcoded path in MsbmsSystInstanceMgr tests
In TestSupport, I had used a hard coded path to clear the migrations.

Not such a big deal but (frankly) I'm copy/pasting in that same code
into other components testing and forgetting to change it.  So make it
dynamic.  Of course, the better answer might be to not engage in that
copy/paste bad habit.... but not today.