Fix MsbmsSystInstanceMgr test_data.eex.sql
The options were not too well defined and are probably still not quite
right (for example where should the defaults be applied?  In this module
or in the app?)  But at least we should make them consistent around the
idea that minimally the Instance Type will establish defaults for the
datastore context connection pool size.  Once we get into using them, it
will become more clear what further action needs to be taken here.