Refactor parameter names to clarify data taken
For some of the functions which take a map parameter holding record
values, there was no designation in the param name that we were really
expecting a parameter object as opposed to, say, the struct that
represents the data.

Let's change those argument names to include the _param suffix to give a
bit of a hint that we're looking for parameters and not a proper data
object of the type we're working with (though to be fair, they'd
probably work as well).