Refactor owner management functions to take name
Previously, in MsbmsSystInstanceMgr, Owner records were modified or
deleted as identified by their id values.  While this is fine, it's
fairly inconsistent with much of the rest of the API which expects an
internal name.

For now, change the id based functions to be internal_name based.  In
most scenarios, we'll have the internal_name any time we have the id and
it should perform roughly equivalently.