Refactor MsbmsSystInstanceMgr doctests_test.exe
In this refactor, we add setup for individual doctests that require
setup.  This is most often the creation of records to support an example
in the documentation, such as documentation for a delete record
function.

This helps keep the doctests 1) subject to testing; and 2) cleaner
because the pre-requisite setup work happens elsewhere.