Refactor & expand MsbmsSystOptions for parsing
Prior to this commit, this component was focused on just reading a TOML
file presumed to carry some flavor of options for system operation.
While this was OK and consistent with the small, reusable component
philosophy, it also was such a small wrapper around the Toml library
that it really didn't make sense to have on it's own.  Also, as we get
into higher levels of the application, we'll need to be opinionated
about what exactly the options file is to the application.  This point
is being reached now.

Rather than considering this a low-level system primitive component, as
was the case before, we now consider this a higher level construct, on
par with Enums, Settings, etc. where the idea of startup parameters and
what they actually are comes from.  With this new thinking, we can now
add parsing functions which extract our specific parameters from the
TOML file.  Naturally, this tightly couples the idea to options to the
higher levels of the application, but that's probably appropriate; other
higher level components such as the Instance Manager and the components
responsible for the Global system management will need to work with
these options so continuing to have a separate component for these
options makes sense and we can expand it if the application has need of
greater information to be stored in the options TOML file.

In general the changes include are:

  * Update the example/testing TOML files to be what our application
    will expect.  Do this across all startup options TOML files,
    including those in other components than MsbmsSystOptions.

  * Add "parsing" for retrieving values from the loaded options
    returning those values in a way that will feel natural for their
    callers; handle any transactions or discard any noise from the
    result.

  * Add MsbmsSystDatastore as a dependency for MsbmsSystOptions.  This
    is only being done because we now produce
    MsbmsSystDatastore.Types.dbserver/0 maps and it's good to make that
    clear in typespecs.  What is unclear is if it is worth the extra
    dependency (and all that comes with it) just for typespecs.  For now
    do it and think about it more later; if we change our mind we can
    always refactor it out.

  * Refactor the docs to match our current standards better, also redo
    some of the doctests to be better fit for example purposes.

  * Refactor testing files to match our current standards better.