Refactor MsbmsSystInstanceMgr.Types Instance types
Updates to the structure of the MsbmsSystInstanceMgr.Data.SysInstances
data structure, particularly those related to the instance_options field
are introduced into the Types module, including supporting types as
needed.

The instance_options field meaning and use floated around for a bit, but
at this time I think it's fairly well settled.  Instance options are now
more tightly coupled to what the MsbmsSystDatastore module needs to know
about database connections, including which server the instance belongs
to, the database contexts which need connecting, and the various "codes"
related to account credential management.