Add instance type functional types to init data
Instance types now require a functional type.  The intent of the new
functional types mostly about filtering which database servers are
available for use with a given instance; database server entries in the
options TOML file designate which instance types (or "instance classes")
are supported by the server, if any.