Refactor database login context privileges
This is a cross-component conceptual refactor of how database privileges
are granted to roles.

Previously, each component's database objects determine what database
roles it was expecting to set privileges for.  The effectively coupled
ideas about what the application needs around database security and
connection barrier were with each individual component (which shouldn't
care about such things).  This was almost exclusively a problem in the
API views of each component.

This change removes the expected application contexts and setting of
login role privileges from the components and pushes that definition
into each component where the opinions on what's required is established
in the first place.  This makes the component decoupled from the
consumer, at least in this regard (I can't think of others however).

While we're here, we fix a privileges bug in the granted privileges for
the msbms_syst.syst_enum_items API view; a bad role assignment led to
failures in later functions which was incorrectly fixed by granting
those failing features SECURITY DEFINER (too loose).  With the correct
error recognized, the root cause is fixed and the prior invalid fix is
reversed.