Cleanup datastore_options related issues
In a variety of modules, inconsistencies concerning how
datastore_options are created have been cropping up as the types have
been refactored over time.

This commit is a cross component effort to clean that up.  We're getting
our targets from Credo, Dialyzer, and the ElixirLS internal Dialyzer
reporting with the goal that all three don't report errors.

Note that differences in `mix dialyzer` and the ElixirLS internal
Dialyzer configuration need to be sussed out; really they should find
and report the same issues because they're both Dialyzer... makes for a
bit of a trust issue.