Refactor MsbmsSystInstanceMgr context management
Prior to this commit, Instance options were kept in a JSON field and
were rather ad hoc.  There were also some defaults for Instance Types
which could be defined and which also validated the Instance options
data.

This was starting to get cumbersome because the data was large enough
and didn't map cleanly to Instance Types when it came to validation and
some definitions.  For example, the required contexts an application
uses to connect to a database are really defined for an application, but
this system required that the required contexts be duplicated for each
instance type; any error in that duplication probably would have been
application breaking for some services.

This commit removes the reliance on the JSON fields and establishes
tables or table columns for context management or defaulting.  This
allows us to manage the data at the correct levels of normalization and
to rely more on database integrity features for a deeper validation than
we should attempt in the application.