Refactor MsbmsSystInstanceMgr test data names
The testing data was not matching places where we're going to be
generating names or was just more cumbersome than it needed to be.

Shorten the names and allow the system to aggregate names for automatic
instance context naming (for example).

Also remove some remaining references to deprecated Instance Type record
functional types (and related data structures).