Fix MsbmsSystDatastore testing cleanup
One of the test phases didn't clean up it's data store on completion
forcing a manual clean up to make it ready for the next test runs.

Add the clean-up.  This still doesn't really conform to our improved
testing processes (at all), but it's sufficient for now.  This debt
seems like it's worth it given the effort to "pay now".