Refactor InstanceTypeContext set function name
The previous name of set_instance_type_context_values was originally
set when the thinking was that there was more than one possible
attribute that could be set for the Instance Type Context record.  Turns
out this isn't the case: there is only one attribute that is
maintainable.

So rename the function to indicate that it's changing that attribute
(default_db_pool_size).

API documentation for the function was also added; there just seems to
be no clean way to exclude it from this commit.