Add table syst_instance_type_applications
While working on the application code related to instance_type_contexts,
it became apparent that the we needed something more general than the
Instance Type/Application Context association in order to appropriately
manage that data.

The msbms_syst_data.syst_instance_type_applications table represent the
idea that certain Instance Types may or may not service a given
application.  That Instance Type Contexts exist is a consequence of the
relationship between an Instance Type and An Application.  With this in
the data model, that relationship becomes explicit vs. implicit from the
existing Instance Type Contexts as was the case previously.  (That and
that previously we assumed all Instance Types would carry Contexts for
all applications... something understood to probably not be true).