Refactor list returning get_* funcs to list_*
So there's been a bit of a naming inconsistency where sometimes I'll use
"get_*" for certain lists of queried objects and "list_*".  It's not
been completely arbitrary.  The rule was if we were returning some sort
of specific selected columns I'd use "list_", but if I was returning
a list of Schema Structs, I'd use "get_*".

I don't think the prior rule was bad, per se, but it did make usage a
bit less obvious.  For example, if I called "get_<something(s)>" I would
know I was dealing with some sort of Schema structs, but I'd have to pay
careful attention to whether or not the get was for a singular object
(get_instance) or for plural objects (get_instances) to know if I was
getting a list or a single object.  Given that the structs are also
named in the plural, even for a singular instance of a struct, I think
things get more potentially confusing.

Make the function names clearer as to what is being returned.  For
lists of things, regardless of their construct, use "list_*" as the name
prefix.  For individual objects, we can "get_*" those.

Apply this to all components existing at commit time.