Refactor MsbmsSystDatastore owner determination
Previously, the database owner was a string at the datastore_options top
level which named the database role that was owner.  This also required
a context with the same database_role name as the owner to exist as a
no-id/login_context = false context so that it would be created.

Simplify that a bit by relying only on the context to determine the
owner.  This way we have a single place and single statement indicating
database ownership and db owner role creation.

Also, update all the users of MsbmsSystDatastore with the new options
structure.  This is mostly (entirely?) found in test prep outside of the
datastore component (at this point).