Add component MsbmsSystUtils; refactor options
This commit adds the MsbmsSystUtils.  Right now it's more boiler plate
than actual logic by a long shot; only one utility function has been
added: resolve_options/2 for resolving function options to defaults if
appropriate.  I expect there to be more functions to fall into this, but
I could be wrong.

Also refactor all function options which were resolving their defaults
on their own.  Doing it this way will save a line or two of code, but
more importantly will make some behaviors more consistent.  For example,
if a parameter is explicitly given as nil, should that be set to the
default value for that parameter or should nil be treated as the final
value?  It answer really could go either way and will sometimes depend
on the specific circumstances; but without a standard it may well vary
in cases that don't care just making things inconsistent. We decide here
to make that nil scenario mean to use the default value if one is
defined; any exceptions can always just not call this function at the
cost of an extra line or two of code.

Also, there's some extraneous rearranging of use/import/alias/require to
appear before the moduledoc stuff where that wasn't true before.  I find
that reading the code it's easier to see the names elements at the top
and I'm not really reading the documentation (since I know it already).