Refactor MsbmsSystInstanceMgr generally
Throw away pretty much everything here.

While working the various data access functions and thinking through a
standard to get them to be consistent across the application, a thought
crossed my mind: is any of this really necessary?  Or is it just more
lines of code chasing an abstraction that doesn't really solve a
problem?

I've clearly concluded that the abstraction isn't necessary.  The
original idea was to hide implementation details, like database access
and structure, behind an abstraction which would allow for greater
separation of concerns across components.  But in the end I don't think
we get there.  For example a user interface will have to, in the various
forms, access most of the data as-is at some point.  With the prior code
this means that the UI forms will still have pretty deep knowledge of
the database and will need to change as the database changes, but now we
have the abstraction code to also maintain in those instances.  For
master data like that in the instances support database, I don't think
we can really create an abstract representation of the data for the
higher levels of the application such that we can decouple those higher
level functions from their lower level implementation details.

Note that this doesn't hold true for MsbmsSystSettings, MsbmsSystEnums,
or even MsbmsSystOptions.  In these cases the implementation details
include mechanisms such as caching and file parsing details which can be
meaningfully placed behind an abstraction.  It's just with master data
and the pre-existing abstraction over data that Ecto brings, at least
here, there's no value to further abstraction.

So throw that abstraction away.  At least for this component.  Most of
these functions would have been consumed in this component anyway.

There are a number of defined types that were there to support those
abstractions.  Those could yet still be useful, so don't get to eager to
throw those away just yet.