Refactor MsbmsSystInstanceMgr schema struct org
Prior to this commit, the Ecto schema structs broke our API vs.
implementation separation standard in that changesets were quite
implementation heavy, but were defined in the struct file itself, which
was meant to also be our API definition.

In this commit, we move those to files under a new Data/Validators
directory.  Files here should be about validation of data.  We also
create another new directory Data/Helpers.  These are functions that are
most likely used in validation, but aren't directly about validation and
conceivably could be used in other contexts.  Internal name to ID
resolution is an example of this possibility.

The name for Validators isn't great and the separation between
Validators and Helpers may be too much abstraction; but right now it was
better than what we had before so long as we're committed to separating
API and implementation... it'll do for now.