Add OTP management features to MsbmsSystDatastore
When originally written, an assumption was made that the repositories
started under an Ecto application related supervision tree and weren't
directly managed by the application.  Turns out this isn't true.  There
is an Ecto Registry for find Repos that require no dependent application
work, but supervision of started processes aren't put in that
supervision tree; in retrospect this isn't surprising since we start the
repo with `start_link/1`, but at the time I didn't think of it that way.

So we add OTP supervision features to the module to ensure that we have
this managed correctly.  Starting a datastore starts a DynamicSupervisor
which will supervise its own Contexts (dynamic repos).  We also add
child_spec and start_link/1 functions so that the datastore supervisor
can be placed in the broader application supervision tree.  We do this
for Datastore Contexts as well since there can be reasons to stop and
start contexts independently.  While the functionality allows contexts
to be placed under arbitrary supervisors, the expectation is that they
will exist under the supervisor established for their owning datastore.

In addition to these changes, many of the privileged and dba function
calls have added code to capture any starting dynamic repo established
for the process and to reset to that repo once the action requiring
dba or privileged access is done.  When this wasn't happening, a call
to one of the these dba/privileged functions would cause the process
dynamic repo to point to the dba/privileged repo and the caller would
have to manage the active repo: the caller should expect things to just
work as though nothing changed.

Some other small/minor clean-up has also taken place, though the new
changes have a fair amount of clean-up to be really complete (typspecs,
docs, etc).