Change MsbmsSystOptions sample TOML db_log_level
This setting was suggesting debug, but this is too verbose and shouldn't
be the value that is de facto recommendation of the testing_options.toml
file.

Show value "warn" instead.  It's possible that "warning" might be better
but right now I can't even tell what the difference would be between
"warn" and "warning".  Either way, this is the most aggressive log level
that should be considered for non-development operations.