Fix bad MsbmsSystDatastore API function signature
start_datastore was sill defined in the MsbmsSystDatastore API module as
being start_datastore/1 when in reality it is now start_datastore/2.

This didn't cause issues because the implementation function incorrectly
defaulted the second argument meaning that the API call would still
function.  In reality that argument default should happen in the API
definition itself (following our common practice).