Refactor database related logging setup
When originally designing the functionality, I erred in thinking that
the log level settings for the database connection were related to
selecting what would be logged and what would not be logged.  In reality
the Ecto related log level settings determine what log level queries
assume.  So, for example, if the setting is log: :info, queries will be
considered `:info` log items rather than being filtered out because
they were a lower level.  The filtering portion happens elsewhere.

So make the following changes so that we get better logging and a more
overall appropriate setup:

  - Remove the MsbmsSystDatastore.Types.db_server() db_log_level
    attribute and all references in places that comply with that type
    definition.

  - Set a standard query log level of "debug" for all datastore
    operations.  This is passed as part of the Ecto dynamic repo
    start_link call.

  - Add config.exs, dev.exs, and test.exs configuration files to all
    components setting dev as log level "debug" and test as "info".

  - Add prod.exs and runtime.exs configuration files to the msbms
    application.  In prod.exs add compile time removal of less than
    info logging (we shouldn't need debug information in production,
    though that's more a bold assumption than a reasoned conclusion).
    Our runtime default will be info.