Refactor typespec of stop_datastore
This was an effort, which failed, to quiet Dialyzer warnings about
invalid typing in some calls for stop_datastore.  To be fair, the update
here is more strictly speaking correct since the full Datastore Context
is not required.

In the end I couldn't solve the Dialyzer complaint (I tried other ways
of altering the typespec, too) and it's not really telling me what it
doesn't like with sufficient specificity to be helpful.  Uncle!  Tests
are passing and until I have reason to do otherwise I'm just going to
ignore Dialyzer.