Refactor datastore_context type definition
There is a fundamental conflict in the typespec definitions for types as
compared to the way the types can get used.  In this case, we used the
type to make a statement about what a complete datastore_context object
looks like; but when it comes to practical usage of this type there are
functions which simply don't need the fully complete record to do their
work.  This means there is no one definition of what a complete
datastore_context is for all usage purposes: it's contextual.  We could
create copies of the types which enforce what each call's expectations
for data requirements are, but that seems wasteful which leaves us with
the alternative approach: simply don't express the abstract idea of a
complete datastore_context using the typespec and instead think of it as
a definition of what's allowed in all contexts.

This is a recurring theme and part of the problem of static typing and
static analysis in a language like Elixir.