Remove Dialyzer from CI build checks
So far Dialyzer has played only a minimal role in improving the
correctness of the application.  Most of the time, the errors it finds
are extremely minor or pedantic errors in specification; given that
the errors are conveyed with a fairly low signal to noise ratio, the
time it takes to resolve the errors compared to the nature and impact of
the error is simply not worth it.  In fact, more often than not the
found errors have no runtime impact and refer more to how correctly I
wrote the typespec.  While having correct typespecs is important, too,
there are limits since they're only real use beyond Dialyzer is
documentation.  If the error found in the typespec doesn't change what
is communicated to the human reader in any substantive way there is a
decent argument for not making our CI testing pass/fail based on it.
The alternative of excluding it everywhere it doesn't work as desired
also seems like an anti-pattern: either it's a useful static analysis
tool or it's a gut check that isn't to be relied upon otherwise.

Or to put it as Chris McCord did on the topic: "Dialyzer as it exists
today is a net negative value add for me. I only use it if forced. In
principle it checks out, but in practice it always costs more than it
saves."

We'll continue to use it, but in the sanity check role as needed.  Maybe
I'll change my mind later if typing bites me, but I think the
combination of documented typespecs, reasonable tests, suitable guards,
and pattern matching will ultimately suffice in preventing most of the
worthwhile errors I'd hope to finding with Dialyzer.