Final push for MsbmsSystInstanceMgr merge
This commit is principally about making the MsbmsSystInstanceMgr
component ready for merging into the next_version branch.  Most of the
changes here are about adding typespecs and documentation were it had
not been included already.  There are some other smaller changes as
well.

  - Add API documentation and organize the documentation via the mix.exs
    configuration points for ExDoc.

  - Add typespecs to public functions that didn't have them defined
    already.

  - Remove the instances_list_item type from MsbmsSystInstanceMgr.Types
    as it's no longer used.

  - A number of functions which should return result tuples didn't
    return an error tuple on failure.  This is corrected, though I think
    we should look at many of these, the root cause error may well be
    opaque more often than it should be.

Not in this commit, but something to think about is documentation for
the structs.  At least a reference to the database documentation is
likely in order.  We'll address when the documentation production is
more formally organized.