Fix MsbmsSystDatastore context connection privs
Datastore contexts were failing to connect in the CI environment due to
lacking CONNECT privileges to the database.  Previously these were
explicitly granted to the context roles and worked, so something
somewhere has changed, but this is likely a better assumption anyway.

Add code to explicitly grant the connect privilege to login database
roles backing Datastore Contexts.

Note that this might not be as robust as we'd like.  Role creation is
transactional for all database roles, but this grant is necessarily
outside of that transaction (the database creation is after the role
creation).  A successful role creation with a failed GRANT operation
would leave things in an inconsistent state.  While unlikely to happen
in practice, we should think about making this more robust.