Add explicit CONNECT revocation on Context drop
The CI test database doesn't want to drop context backing roles because
of the CONNECT privilege we added earlier (because the CI test database
didn't want to let context backing roles connect).  So tests fails and
contexts are not dropped.

Add functionality to drop context CONNECT privileges on the datastore
database.  Note that, similar to the CONNECT privilege grant functions,
this currently feels hacky.  I think there is probably a better, more
resilient solution, but it's more important to get something that works
sufficiently than something that works perfectly or elegantly right now.