Fix timeouts when stopping Datastore Context
For some reason we get different error messages whenever we use a
timeout we get errors trying to stop the dynamic repos... sometimes.
More than that, the nature of the errors are different if it's run
through Erlang 24 or 25... still errors, but different.

Having an infinite timeout is not good, but it's in the near term it's
not critical path from a project perspective.  So note it as a TODO and
move on.