Fix sequencing in stop_instance/2
We were seeing errors shutting down database connections and the
thinking was that stopping the supervisor was leaving things dangling
leading to later test errors trying to drop the database (still
connected).

This commit changes the sequence to stop the datastore contexts first,
and then stop the instance supervisor.

In retrospect the issue was likely something else altogether.  The order
shouldn't matter and the instance supervisor shutdown should take things
down correctly.  Anyway, this is more technically correct so we'll keep
it.