Refactor owner_test.exs to make tests more dynamic
We really shouldn't have very tailored, predictable tests.  While not
exactly property testing, at least pull a database order record from the
database to try.

The only downside to this is if a concurrent test changes/removes the
record midway through our tests in this commit, we could get extraneous
or intermittent errors.