Remove syst_access_account global login constraint
On reflection, owned accounts should not necessarily be restricted from
global login features.

The global login feature was conceived around independent consultant/
accountant use cases, where the independent contract may service many
different Owners and so would correctly be authenticated once and then
allowed to select which Owner and Instance to access from a global
login process; this is opposed to a login user interface accessed at an
Owner specific URL (for example) which only can access the specific
Owner's Instances by definition.

If we think about a more extended use case however, such a contractor
doesn't have to be an individual: the user may be an employee of a
company that services other Owners.  In this case, the consultant/
accountant/contractor user could be an owned account as they are simply
the representative of a service provider, but the global login option
may still be the best presentation for them if they are in other
companies instance more often than their own.  There is no compelling
reason I can think of right now why not to allow owned accounts to use
the global authentication process.

Remove the constraint and better document the rational of global access.