Add third_party/toml-elixir
This is an updated toml-elixir library which includes the mixed-type
arrays allowed by TOML 1.0.

The changes here have been submitted to upstream but the library is not
frequently maintained so it may wait there for some time.

Once our changes are merged upstream, then we can remove this third
party dependency (unless we decide to vendor all external dependencies.)