Add MsbmsSystDatastore nested build plans feature
The component based approach that MSBMS is using has proven very helpful
in defining and enforcing good boundaries between concerns, but with the
downside that there's a fair amount of boilerplate.  This was also true
for our migrations and especially component testing migrations.  The
testing build plans for each component have to include up-to-date
file lists for each of its dependencies which is difficult to maintain
when there are refactorings of the dependency.

This change allows each component to define its own build plan which can
then be referenced in the build plan of any dependent.  So when a
dependency is refactored and its build plan changes, it only has to be
changed in one place.  Naturally the success of this approach depends
heavily on ensuring that the dependency graph between components is
unidirectional (circular dependencies not allowed).