Remove msbms_syst_schemata.eex.sql appusr privs
This file is now a general, cross component sub-plan; not the component
specific testing setup it use to be.  So remove the grants to users
since this file may be loaded into environments where that context
doesn't exist.

This does mean that these permissions need to be added back in the
component testing SQL files and other more application specific
locations.