Reorder MSBMS Sublime project folders
This better reflects the dependency graph between components, with
directories higher on the list being more basic/fundamental.

Reminder that circular dependencies are forbidden which has the effect
of evolving the dependency graph in one consistent direction.  In this
case, lower listed components may depend on higher listed components,
but not the reverse (this holds true for the top level Elixir
components; sub-directories are probably sorted in dictionary order).