Remove MsbmsSystMnesia
On reflection, this really shouldn't be a component.  The purposes of
creating a bespoke component was pretty thinly justified and even then
just not correct; for example, being sure we had a Mnesia schema created
that then dependents could just use prematurely forces or complicates
the process of choosing a location for the Mnesia data directory or
specification of the nodes.  These are really top level, Application as
application (not Application as component) concerns.  Beyond that we
were just going to wrap an API which doesn't really give us any
worthwhile benefit.