Refactor toml dependency to be git based
We're currently using a branch of the toml library that hasn't yet been
incorporated into the upstream library yet (it's been submitted as a
PR).  I tried making the dependency a path dependency and using a
git subtree to link the repo/branch to the musebms project.  That
appears to have caused more trouble than its worth.

Making the dependency a git dependency should bypass the git headaches
and be most compatible with the hex based libraries.