Refactor SystGlobalPasswordRules insert/delete
As things evolved, the global password rules became the base level for
system wide minimum password standards.  Given the structure of the
table, this means that there should always be exactly one record in the
syst_global_password_rules table since the one record contains all the
values needed for the supported password rules.

Remove the ability to insert or delete this rule via the API views and
add a starting point for the values as seed_data.