Refactor SystCredentials for multi-factor auth
There are two additional credential types needed to fully support MFA
via TOTP.  First, we need to have TOTP recovery codes in case of TOTP
authenticator device loss.  For this we just need a specialized
credential type.

Next is support for the common practice of remembering a user's devices
as being trusted for purposes of MFA; usually this is for a period of
time such as 30 days where the user is not prompted for the MFA
authenticator code.  We can support this via the credentials table, too,
so long as we have a way to display the known hosts to the relevant user
so that they can choose to forget a specific device.  So add an optional
`external_name` field to syst_credentials for this purpose (named along
the lines of other uses of `external_name`.