Refactor network rule validation function
In SystGlobalNetworkRules, SystOwnerNetworkRules, and
SystInstanceNetworkRules, Credo was complaining about the form of the
validate_cidr_or_host/1 private function; it basically said cyclomatic
complexity was too high (12 checked against max 9).  So refactor in a
way to silence the warnings.

Not sure this is a good change.  While in some ways it's definitely more
idiomatic Elixir, I'm not sure it's really less complex.  To be fair it
is easier to read the conditional expressions where they set variable
values rather than exist directly in the conditional expression, but
having to scan the now various functions and function heads in a non-
linear fashion also doesn't seem perfect.

We'll run for now so that this will pass CI (without manual exceptions
to the Credo check).